+++ app/src/main/java/ru/mertsalovda/calculatorofpower/App.kt
@@ -2,7 +2,7 @@ package ru.mertsalovda.calculatorofpower
 
 import android.app.Application
 import ru.mertsalovda.core_api.mediator.AppWithFacade
-import ru.mertsalovda.core_api.mediator.ProvidersFacade
+import ru.mertsalovda.core_api.providers.ProvidersFacade
 
 class App : Application(), AppWithFacade {
