+++ app/src/main/java/ru/mertsalovda/calculatorofpower/AppComponent.kt
@@ -4,7 +4,7 @@ import android.app.Application
 import android.content.Context
 import dagger.BindsInstance
 import dagger.Component
-import ru.mertsalovda.core_api.mediator.AppProvider
+import ru.mertsalovda.core_api.providers.AppProvider
 import javax.inject.Singleton
 
 @Singleton
