+++ app/src/main/java/ru/mertsalovda/calculatorofpower/FacadeComponent.kt
@@ -4,8 +4,8 @@ import android.app.Application
 import dagger.Component
 import ru.mertsalovda.core.CoreProvidersFactory
 import ru.mertsalovda.core_api.database.DatabaseProvider
-import ru.mertsalovda.core_api.mediator.AppProvider
-import ru.mertsalovda.core_api.mediator.ProvidersFacade
+import ru.mertsalovda.core_api.providers.AppProvider
+import ru.mertsalovda.core_api.providers.ProvidersFacade
 
 @Component(
     dependencies = [AppProvider::class, DatabaseProvider::class],
