+++ sources/core/src/main/java/ru/mertsalovda/core/CoreProvidersFactory.kt
@@ -1,7 +1,7 @@
 package ru.mertsalovda.core
 
 import ru.mertsalovda.core_api.database.DatabaseProvider
-import ru.mertsalovda.core_api.mediator.AppProvider
+import ru.mertsalovda.core_api.providers.AppProvider
 import ru.mertsalovda.core_impl.DaggerDatabaseComponent
 
 object CoreProvidersFactory {
