--- sources/core_api/src/main/java/ru/mertsalovda/core_api/mediator/MediatorsProvider.kt
@@ -1,12 +1,0 @@
-package ru.mertsalovda.core_api.mediator
-
-interface MediatorsProvider {
-
-    fun provideMainMediator(): MainMediator
-
-    fun provideBasicMediator(): BasicCalculatorMediator
-
-    fun provideScientificMediator(): ScientificCalculatorMediator
-
-    fun provideConverterMediator(): ConverterMediator
-}
