+++ sources/core_impl/src/main/java/ru/mertsalovda/core_impl/DatabaseComponent.kt
@@ -2,7 +2,7 @@ package ru.mertsalovda.core_impl
 
 import dagger.Component
 import ru.mertsalovda.core_api.database.DatabaseProvider
-import ru.mertsalovda.core_api.mediator.AppProvider
+import ru.mertsalovda.core_api.providers.AppProvider
 import javax.inject.Singleton
 
 @Singleton
