+++ sources/feature-converter/src/main/res/values/strings.xml
@@ -5,4 +5,6 @@
     <string name="speed">Скорость</string>
     <string name="weight">Масса</string>
     <string name="area">Площадь</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
